+++ nens_auth_client/conf.py
@@ -18,6 +18,7 @@ class NensAuthClientAppConf(AppConf):
 
     INVITATION_EMAIL_SUBJECT = "Invitation"
     INVITATION_EXPIRY_DAYS = 14  # change this to change the default expiry
+    USERNAME_MAX_LENGTH = 50
 
     ERROR_USER_DOES_NOT_EXIST = "No user account available for these credentials."
     ERROR_USER_INACTIVE = "This account was set to inactive."